Refactor testing framework
