+++ .idea/runConfigurations/Visual_Tests.xml
@@ -1,7 +1,8 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Visual Tests" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
-    <option name="command" value="test --profile=&quot;desktop&quot; --features=&quot;vis_test&quot;" />
+    <option name="command" value="test --workspace --profile=&quot;desktop&quot; --features=&quot;vis_test&quot;" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$/rs" />
+    <envs />
     <option name="emulateTerminal" value="true" />
     <option name="channel" value="DEFAULT" />
     <option name="requiredFeatures" value="true" />
@@ -9,7 +10,6 @@
     <option name="withSudo" value="false" />
     <option name="buildTarget" value="REMOTE" />
     <option name="backtrace" value="SHORT" />
-    <envs />
     <option name="isRedirectInput" value="false" />
     <option name="redirectInputPath" value="" />
     <method v="2" />
